refactor the pinout SVG generation logic for improved code organization and readability
