+++ lib/pinout/convert-circuit-json-to-pinout-svg.ts
@@ -16,6 +16,11 @@ import { createSvgObjectsFromPinoutSmtPad } from "./svg-object-fns/create-svg-ob
 import { createSvgObjectsFromPinoutPort } from "./svg-object-fns/create-svg-objects-from-pinout-port"
 import { getSoftwareUsedString } from "../utils/get-software-used-string"
 import { CIRCUIT_TO_SVG_VERSION } from "../package-version"
+import {
+  calculateLabelPositions,
+  type LabelPosition,
+} from "./calculate-label-positions"
+import { getClosestEdge } from "./pinout-utils"
 
 const OBJECT_ORDER: AnyCircuitElement["type"][] = [
   "pcb_board",
@@ -36,69 +41,7 @@ export interface PinoutSvgContext {
   transform: Matrix
   soup: AnyCircuitElement[]
   board_bounds: { minX: number; minY: number; maxX: number; maxY: number }
-  label_positions: Map<
-    string,
-    {
-      text: string
-      aliases: string[]
-      elbow_end: { x: number; y: number }
-      label_pos: { x: number; y: number }
-      edge: "left" | "right" | "top" | "bottom"
-    }
-  >
-}
-
-function getPortLabelInfo(
-  port: PcbPort,
-  soup: AnyCircuitElement[],
-): { text: string; aliases: string[] } | null {
-  const source_port = su(soup).source_port.get(port.source_port_id)
-  if (!source_port) return null
-
-  const eligible_hints =
-    source_port.port_hints?.filter(
-      (h) =>
-        !/^\d+$/.test(h) && !["left", "right", "top", "bottom"].includes(h),
-    ) ?? []
-
-  let label = eligible_hints[0]
-  if (!label) label = source_port.name
-
-  if (!label) return null
-
-  const aliases = eligible_hints.filter((h) => h !== label)
-
-  return { text: label, aliases }
-}
-
-function getClosestEdge(
-  port_pos_real: { x: number; y: number },
-  board_bounds: { minX: number; minY: number; maxX: number; maxY: number },
-): "left" | "right" | "top" | "bottom" {
-  const dists = {
-    left: port_pos_real.x - board_bounds.minX,
-    right: board_bounds.maxX - port_pos_real.x,
-    top: board_bounds.maxY - port_pos_real.y,
-    bottom: port_pos_real.y - board_bounds.minY,
-  }
-
-  let closest_edge: "left" | "right" | "top" | "bottom" = "left"
-  let min_dist = dists.left
-
-  if (dists.right < min_dist) {
-    min_dist = dists.right
-    closest_edge = "right"
-  }
-  if (dists.top < min_dist) {
-    min_dist = dists.top
-    closest_edge = "top"
-  }
-  if (dists.bottom < min_dist) {
-    min_dist = dists.bottom
-    closest_edge = "bottom"
-  }
-
-  return closest_edge
+  label_positions: Map<string, LabelPosition>
 }
 
 export function convertCircuitJsonToPinoutSvg(
@@ -178,249 +121,14 @@ export function convertCircuitJsonToPinoutSvg(
     ports_by_edge[edge].push(port)
   }
 
-  const label_positions = new Map<
-    string,
-    {
-      text: string
-      aliases: string[]
-      elbow_end: { x: number; y: number }
-      label_pos: { x: number; y: number }
-      edge: "left" | "right" | "top" | "bottom"
-    }
-  >()
-
-  const V_SPACING = 20
-
-  const STAGGER_OFFSET_MIN = 20
-  const STAGGER_OFFSET_PER_PIN = 2
-  const STAGGER_OFFSET_STEP = 15
-  const ALIGNED_OFFSET_MARGIN = 10 // Margin beyond the last staggered point
-
-  const FONT_SIZE = 11
-  const BG_PADDING = 5
-  const LABEL_RECT_HEIGHT = FONT_SIZE + 2 * BG_PADDING
-  const LABEL_MARGIN = 5
-
-  // Left edge
-  const left_ports = ports_by_edge.left
-    .map((port) => ({
-      port,
-      y: applyToPoint(transform, [port.x, port.y])[1],
-      label_info: getPortLabelInfo(port, soup),
-    }))
-    .filter((p) => p.label_info)
-    .sort((a, b) => a.y - b.y)
-
-  if (left_ports.length > 0) {
-    const board_left_x = applyToPoint(transform, [board_bounds.minX, 0])[0]
-
-    const num_labels = left_ports.length
-    const middle_index = (num_labels - 1) / 2
-
-    const stagger_offset_base =
-      STAGGER_OFFSET_MIN + num_labels * STAGGER_OFFSET_PER_PIN
-
-    const max_stagger_offset =
-      stagger_offset_base + middle_index * STAGGER_OFFSET_STEP
-    const aligned_label_offset = max_stagger_offset + ALIGNED_OFFSET_MARGIN
-
-    const total_labels_height =
-      num_labels * LABEL_RECT_HEIGHT +
-      Math.max(0, num_labels - 1) * LABEL_MARGIN
-    let current_y =
-      (svgHeight - total_labels_height) / 2 + LABEL_RECT_HEIGHT / 2
-
-    left_ports.forEach(({ port, label_info }, i) => {
-      const dist_from_middle = Math.abs(i - middle_index)
-      const stagger_rank = middle_index - dist_from_middle
-      const stagger_offset =
-        stagger_offset_base + stagger_rank * STAGGER_OFFSET_STEP
-
-      const elbow_end = { x: board_left_x - stagger_offset, y: current_y }
-      const label_pos = { x: board_left_x - aligned_label_offset, y: current_y }
-
-      label_positions.set(port.pcb_port_id, {
-        text: label_info!.text,
-        aliases: label_info!.aliases,
-        elbow_end,
-        label_pos,
-        edge: "left",
-      })
-      current_y += LABEL_RECT_HEIGHT + LABEL_MARGIN
-    })
-  }
-
-  // Right edge
-  const right_ports = ports_by_edge.right
-    .map((port) => ({
-      port,
-      y: applyToPoint(transform, [port.x, port.y])[1],
-      label_info: getPortLabelInfo(port, soup),
-    }))
-    .filter((p) => p.label_info)
-    .sort((a, b) => a.y - b.y)
-
-  if (right_ports.length > 0) {
-    const board_right_x = applyToPoint(transform, [board_bounds.maxX, 0])[0]
-
-    const num_labels = right_ports.length
-    const middle_index = (num_labels - 1) / 2
-
-    const stagger_offset_base =
-      STAGGER_OFFSET_MIN + num_labels * STAGGER_OFFSET_PER_PIN
-
-    const max_stagger_offset =
-      stagger_offset_base + middle_index * STAGGER_OFFSET_STEP
-    const aligned_label_offset = max_stagger_offset + ALIGNED_OFFSET_MARGIN
-
-    const total_labels_height =
-      num_labels * LABEL_RECT_HEIGHT +
-      Math.max(0, num_labels - 1) * LABEL_MARGIN
-    let current_y =
-      (svgHeight - total_labels_height) / 2 + LABEL_RECT_HEIGHT / 2
-
-    right_ports.forEach(({ port, label_info }, i) => {
-      const dist_from_middle = Math.abs(i - middle_index)
-      const stagger_rank = middle_index - dist_from_middle
-      const stagger_offset =
-        stagger_offset_base + stagger_rank * STAGGER_OFFSET_STEP
-
-      const elbow_end = { x: board_right_x + stagger_offset, y: current_y }
-      const label_pos = {
-        x: board_right_x + aligned_label_offset,
-        y: current_y,
-      }
-
-      label_positions.set(port.pcb_port_id, {
-        text: label_info!.text,
-        aliases: label_info!.aliases,
-        elbow_end,
-        label_pos,
-        edge: "right",
-      })
-      current_y += LABEL_RECT_HEIGHT + LABEL_MARGIN
-    })
-  }
-
-  // Top edge
-  const top_ports = ports_by_edge.top
-    .map((port) => ({
-      port,
-      x: applyToPoint(transform, [port.x, port.y])[0],
-      label_info: getPortLabelInfo(port, soup),
-    }))
-    .filter((p) => p.label_info)
-    .sort((a, b) => a.x - b.x)
-
-  if (top_ports.length > 0) {
-    const board_top_y = applyToPoint(transform, [0, board_bounds.maxY])[1]
-
-    const labels_with_widths = top_ports.map((p) => {
-      const label = [p.label_info!.text, ...p.label_info!.aliases].join(" | ")
-      const textWidth = label.length * FONT_SIZE * 0.6
-      const rectWidth = textWidth + 2 * BG_PADDING
-      return { ...p, rectWidth }
-    })
-
-    const num_labels = labels_with_widths.length
-    const middle_index = (num_labels - 1) / 2
-
-    const stagger_offset_base =
-      STAGGER_OFFSET_MIN + num_labels * STAGGER_OFFSET_PER_PIN
-
-    const max_stagger_offset =
-      stagger_offset_base + middle_index * STAGGER_OFFSET_STEP
-    const aligned_label_offset = max_stagger_offset + ALIGNED_OFFSET_MARGIN
-
-    const total_labels_width =
-      labels_with_widths.reduce((sum, l) => sum + l.rectWidth, 0) +
-      Math.max(0, num_labels - 1) * LABEL_MARGIN
-    let current_x = (svgWidth - total_labels_width) / 2
-
-    labels_with_widths.forEach(({ port, label_info, rectWidth }, i) => {
-      const dist_from_middle = Math.abs(i - middle_index)
-      const stagger_rank = middle_index - dist_from_middle
-      const stagger_offset =
-        stagger_offset_base + stagger_rank * STAGGER_OFFSET_STEP
-
-      const label_center_x = current_x + rectWidth / 2
-      const elbow_end = { x: label_center_x, y: board_top_y - stagger_offset }
-      const label_pos = {
-        x: label_center_x,
-        y: board_top_y - aligned_label_offset,
-      }
-
-      label_positions.set(port.pcb_port_id, {
-        text: label_info!.text,
-        aliases: label_info!.aliases,
-        elbow_end,
-        label_pos,
-        edge: "top",
-      })
-      current_x += rectWidth + LABEL_MARGIN
-    })
-  }
-
-  // Bottom edge
-  const bottom_ports = ports_by_edge.bottom
-    .map((port) => ({
-      port,
-      x: applyToPoint(transform, [port.x, port.y])[0],
-      label_info: getPortLabelInfo(port, soup),
-    }))
-    .filter((p) => p.label_info)
-    .sort((a, b) => a.x - b.x)
-
-  if (bottom_ports.length > 0) {
-    const board_bottom_y = applyToPoint(transform, [0, board_bounds.minY])[1]
-
-    const labels_with_widths = bottom_ports.map((p) => {
-      const label = [p.label_info!.text, ...p.label_info!.aliases].join(" | ")
-      const textWidth = label.length * FONT_SIZE * 0.6
-      const rectWidth = textWidth + 2 * BG_PADDING
-      return { ...p, rectWidth }
-    })
-
-    const num_labels = labels_with_widths.length
-    const middle_index = (num_labels - 1) / 2
-
-    const stagger_offset_base =
-      STAGGER_OFFSET_MIN + num_labels * STAGGER_OFFSET_PER_PIN
-
-    const max_stagger_offset =
-      stagger_offset_base + middle_index * STAGGER_OFFSET_STEP
-    const aligned_label_offset = max_stagger_offset + ALIGNED_OFFSET_MARGIN
-
-    const total_labels_width =
-      labels_with_widths.reduce((sum, l) => sum + l.rectWidth, 0) +
-      Math.max(0, num_labels - 1) * LABEL_MARGIN
-    let current_x = (svgWidth - total_labels_width) / 2
-
-    labels_with_widths.forEach(({ port, label_info, rectWidth }, i) => {
-      const dist_from_middle = Math.abs(i - middle_index)
-      const stagger_rank = middle_index - dist_from_middle
-      const stagger_offset =
-        stagger_offset_base + stagger_rank * STAGGER_OFFSET_STEP
-
-      const label_center_x = current_x + rectWidth / 2
-      const elbow_end = {
-        x: label_center_x,
-        y: board_bottom_y + stagger_offset,
-      }
-      const label_pos = {
-        x: label_center_x,
-        y: board_bottom_y + aligned_label_offset,
-      }
-      label_positions.set(port.pcb_port_id, {
-        text: label_info!.text,
-        aliases: label_info!.aliases,
-        elbow_end,
-        label_pos,
-        edge: "bottom",
-      })
-      current_x += rectWidth + LABEL_MARGIN
-    })
-  }
+  const label_positions = calculateLabelPositions({
+    ports_by_edge,
+    transform,
+    soup,
+    board_bounds,
+    svgWidth,
+    svgHeight,
+  })
 
   const ctx: PinoutSvgContext = {
     transform,
